Algo está dando errado. Descobrir e arruamr :(

+++ src/t3jb/GenTree.java
@@ -7,9 +7,9 @@ public class GenTree {
   private static class TreeNode {
     Node children;
     String name;
-    double data;
+    int data;
 
-    TreeNode( String newName, double newData ) {
+    TreeNode( String newName, int newData ) {
       data = newData;
       name = newName;
       children = null;
@@ -37,7 +37,7 @@ public class GenTree {
       return t.root;
   }
   
-  public void setValue(TreeNode n, double value){
+  public void setValue(TreeNode n, int value){
       n.data = value;
   }
   
@@ -63,7 +63,7 @@ public class GenTree {
       listaFilhos.next = new Node (n);
   }
 
-  public GenTree( String s, double n ) { root = new TreeNode( s, n ); }
+  public GenTree( String s, int n ) { root = new TreeNode( s, n ); }
 
   private TreeNode find( TreeNode n, String val ) {
     if ( n == null ) return null;
@@ -77,7 +77,7 @@ public class GenTree {
     return null;
   }
 
-  public void insert( String p, String s, double f) {
+  public void insert( String p, String s, int f) {
     TreeNode n = find( root, p );
     if ( n == null ) return;
     n.children = append( n.children, new TreeNode( s, f ) );
@@ -146,7 +146,7 @@ public class GenTree {
   
   private TreeNode getBigProbability(TreeNode n){
     if(n.children == null) return n;
-    double percent = 0;
+    int percent = 0;
     TreeNode chosen = n;
     Node f = n.Children();
     while(f!=null){
@@ -169,8 +169,8 @@ public class GenTree {
       }
   }
   
-  private void arrumaValores(TreeNode n, double value){
-      n.data = (double)(value*((float)n.data/100.0f));
+  private void arrumaValores(TreeNode n, int value){
+      n.data = (int)(value*((float)n.data/100.0f));
       Node f = n.children;
       while(f != null){
           arrumaValores(f.child,n.data);
